move mock data to .stories
