--- packages/ui/src/components/comment/comment.stories.svelte
@@ -1,29 +1,0 @@
-<script context="module" lang="ts">
-  import type { Meta } from '@storybook/svelte'
-
-  import { Comment } from './index.js'
-
-  export const meta = {
-    title: 'Molecule/Comment',
-    component: Comment,
-    tags: ['autodocs'],
-  } satisfies Meta<Comment>
-</script>
-
-<script lang="ts">
-  import { Story, Template } from '@storybook/addon-svelte-csf'
-</script>
-
-<Template let:args>
-  <Comment {...args} />
-</Template>
-
-<Story
-  name="Mobile"
-  parameters="{{ viewport: { defaultViewport: 'mobile2' } }}"
-/>
-
-<Story
-  name="Desktop"
-  parameters="{{ viewport: { defaultViewport: 'tablet' } }}"
-/>
