+++ packages/ui/src/components/molecule/comment/comment.stories.svelte
@@ -1,0 +1,45 @@
+<script module lang="ts">
+  import { defineMeta } from '@storybook/addon-svelte-csf'
+
+  import { Comment } from './index.js'
+
+  const { Story } = defineMeta<typeof Comment>({
+    title: 'Molecule/Comment',
+    component: Comment,
+    tags: ['autodocs'],
+    argTypes: {
+      content: {
+        control: 'text',
+      },
+      semester: {
+        control: 'text',
+      },
+      rating: {
+        control: 'number',
+      },
+      likesCount: {
+        control: 'number',
+      },
+      dislikesCount: {
+        control: 'number',
+      },
+    },
+  })
+</script>
+
+<Story
+  name="Default"
+  args={{
+    content: `ส่วนตัวคิดว่าค่อนข้างสบาย
+มีงานสัปดาห์ละครั้ง ปกติจะมีวิดีโอให้ดู ไม่ยาวมาก แต่จะไม่ดูก็ได้ เพราะปกติเราแค่อ่านตัวอย่างที่ให้มาแล้วก็ลองเขียนเลย งานส่วนใหญ่จะให้ 
+Topic กว้างๆมา ถ้าอาจารย์ประจำ sec ไม่ strict มาก ก็เขียนตามใจไปเลย แค่เอาเรื่องที่เรียนในสัปดาห์นั้นมาปรับๆนิดหน่อย 
+ปกติจะไม่อยากเสียเวลามาก ก็เขียนครั้งเดียวแล้วส่งไปเลย เลยเสียเวลาแค่ 1-2 ชม. ต่อสัปดาห์ แต่ถ้าใครจริงจังกับงานเขียนหน่อยก็อาจใช้เวลาเพิ่มขึ้น
+สอบเหมือนกับงานที่ทำ จะไม่อ่านไปสอบก็ได้ถ้าจำพวกงานที่ทำส่งได้ แต่ก่อนสอบมีคลิปรีวิวให้ดูก่อน พึ่งตัวนี้เอาก็ได้
+อันนี้ก็ขึ้นอยู่กับ sec ของเราอาจารย์ให้คะแนนงานค่อนข้างง่าย แต่จะให้คะแนนสอบยากหน่อย แต่เฉลี่ยๆรวมๆกันก็ได้ A
+เพราะเขียนตามใจตัวเองเป็นหลัก ไม่ค่อยยึดกับเรื่องที่เรียนมาเลยรู้สึกไม่ค่อยได้อะไรใหม่ ที่ช่วยน่าจะเป็นการจัดระเบียบความคิดและสรุปให้อยู่ในย่อหน้า`,
+    semester: `ภาคต้น 2565`,
+    rating: 3.5,
+    likesCount: 100,
+    dislikesCount: 0,
+  }}
+/>
